cleaning

+++ ft_memccpy.c
@@ -6,7 +6,7 @@
 /*   By: avarnier <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/04 15:13:28 by avarnier          #+#    #+#             */
-/*   Updated: 2019/11/22 16:25:36 by avarnier         ###   ########.fr       */
+/*   Updated: 2019/12/01 10:39:44 by avarnier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memccpy(void *dst, const void *src, int c, size_t len)
 {
-	int	x;
+	size_t	x;
 
 	x = 0;
 	while (x < len)
