+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: avarnier <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/05 19:29:15 by avarnier          #+#    #+#             */
-/*   Updated: 2019/11/22 18:18:34 by avarnier         ###   ########.fr       */
+/*   Updated: 2019/12/01 10:41:57 by avarnier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t len)
 {
-	int x;
+	size_t x;
 
 	x = 0;
 	while (x < len && s1[x] == s2[x] && s1[x] != '\0' && s2[x] != '\0')
