+++ ft_strtrim.c
@@ -6,16 +6,16 @@
 /*   By: avarnier <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/14 14:45:28 by avarnier          #+#    #+#             */
-/*   Updated: 2019/11/25 17:54:56 by avarnier         ###   ########.fr       */
+/*   Updated: 2019/12/01 10:46:20 by avarnier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int	ft_start(const char *s1, const char *set)
+static size_t	ft_start(const char *s1, const char *set)
 {
-	int x;
-	int y;
+	size_t x;
+	size_t y;
 
 	x = 0;
 	y = 0;
@@ -52,7 +52,7 @@ static int	ft_end(const char *s1, const char *set)
 	return (x);
 }
 
-char			*ft_strtrim(const char *s1, const char *set)
+char		*ft_strtrim(const char *s1, const char *set)
 {
 	if (ft_start(s1, set) == ft_strlen(s1))
 		return (ft_strdup(""));
